fix: avoid panics when Unwrap/Cause return nil.
The std lib's error package says that Unwrap may return nil when
the error doesn't wrap another error; I'm assuming Cause works the
same.

This change fixes a code path where a nil cause was being de-referenced
and causing a "PANIC_DISPLAYING_ERROR" error to be printed instead
of the error we're interested in.